Bug fixes and optimizations for streams.
Streams should use the user's buffer if the caller requested a large
amount of data. Also it turns out the m4 tests were using fseek to
determine if the underlying file was seekable. This means our clever work
of seeking within the buffer if possible was no good, since it never hit
the system call. Removing that case simplified the stream handling a bit.

Also try to be more compliant in the library with a case where O_RDONLY is
zero. It's not on our system, but is on others.