Removed KD's spurious freeze accounting.
On a debug exception, the breaking core needs to freeze the other cores. On
x86, non-maskable NMIs are used to signal the other cores. On ARM, maskable
IPIs are used. As a result, cores may miss the freeze request on ARM. This
is especially bad when the core is also waiting to break into the debugger.
To handle this, certain debug paths must manually check for a freeze quest
and then be sure to ignore a freeze IPI once interrupts are enabled.

Before this change, the code tried to count how many IPIs must be ignored,
but this proved difficult and was not consistent across interrupt
controllers (i.e. the GIC pends an IPI for each unique line, source
source processor and target processor tuple whereas the BCM2709 pends one
IPI per line). The end result of the bad accounting were cores that did
not freeze or cores that hit an assert when they should not have.

The solution here is to just ignore all freeze IPIs on ARM if there is no
debugger freeze owner.