Fixed KD's profiler data status collection.
KD used to pass the profiler data status flags down into the exception
handler, but this can result in a race between cores. If two cores read that
there is memory profiling data to send and then decided to break into the
debugger, only one will end up reading the valid data to send. The other
will read invalid data and send it anyway. The fix is to resample the
profiling data status after breaking into the debugger and only proceeding
to send data if valid data is available.