Improved stack profiling throughput.
Before this change, stack profiling sending data to the debug host every
250ms. On most serial debuggers, it takes ~175ms per-core to send the data,
which eats of 70% of the 250ms period. So, the first improvement is to only
send the data every 500ms. The per-processor buffers are big enough so that
data should not be lost.

The ~175ms per-core gets worse on a multi-core system, because KD freezes the
other cores while sending profiling data. So, on a 4-core system, each core
is really stuck sending data for ~700ms for every send interval. This can
cause threads to get completely starved out by the profiler. The second
improvement in this change is to not require KD to freeze all cores when it
is only sending profiling data. The cores will queue up on KD's debugger
lock, but once a core sends its data it does not have to wait for the other
cores to send their data as well. This goes a long way to prevent thread
starvation.