Fix for regular expression execution.
Fixed a bug in the logic that attempted to restore subexpression match
values upon a mismatch.

The old code's strategy was to go back one element, then traverse up the
regex tree until the choice being modified was found. Along the way, any
subexpressions would be reset. It asserted that if it found any
subexpression, it had better match the current choice being revised. That
assert fired.

The problem is in its code for "going back one" from the element that just
mismatched. It went to the parent's previous child, and then down the
rightmost child all the way. The problem is that branches "|" may not have
taken the rightmost choice, so then while going up the tree a
subexpression may be found that was never part of this set of matches.

Since subexpressions always create a choice in the choice tree, and the
choice being revised is always the most recent one, the fix just undoes the
subexpression that the revised choice points to. There's no need to clamor
around the tree looking for other hidden subexpressions, because every
subexpression has a node in the choice tree.