Faster but deviant regex execution.
The old regex code was overdoing it looking for a match for *. This
change backtracks a lot faster, but might settle on a local maximum for
greedy * when a larger global maximum was available. This may cause some
compatibility problems for those expecting a slightly different set of
regex subexpression regions.

This change may need to be backed out. Ultimately the right solution would
be to reimplement regular expressions using NFA/DFA state machine
construction (ie Thompson-McNaughton-Yamada NFA and powerset construction
to get to a DFA). That would probably be the most compatible.