Added support for absolute timers.
An absolute timer supports generating interrupts based on an absolute timer
value. Typically the hardware will program the desired value into a compare
register and the interrupt will fire when the counter register is greater
than or equal to the compare register. As this does not require changes to
the timer's value, absolute timers can be shared between component requiring
interrupts (e.g. the clock timer) and multiple components that do not
require interrupts (e.g. the time counter).

The ARM Generic Timer is a good example of an absolute timer. Unfortunately,
it does not support periodic modes. That is, an interval cannot be
programmed into the timer to cause repeated interrupts. As a result, this
change adds higher level support to convert non-periodic, absolute timers
into periodic timers. This requires converting the periodic interval into
an absolute time during arm requests and reprogramming the timer with a new
absolute time during acknowledge interrupt. In order to synchronize the
rearming performed by acknowledge interrupt with new calls to arm or disarm,
all action for these types of timers must be taken at the interrupt run
level.