Removed private hardware module MPIDR routines.
Both the GIC and the Broadcom interrupt controller need the MPIDR value in
in a few locations. Before the kernel services interface was removed, these
modules implemented their own assembly routine to get the MPIDR. This change
removes those files in favor of using the shared Ar version of the routine.