Separated the clock IPI from the clock interrupt.
Prior to this change, software generated clock IPIs and hardware generated
clock interrupts were sharing the same vector, KINTERRUPT, and the same
interrupt handler. This had two downsides. First, the hardware module was
being invoked to acknowledge the clock interrupt on IPIs, when there was no
interrupt to acknowledge. Second, when the clock owner (P0) receives on IPI
on non per-processor clocks, it was needlessly re-broadcasting that IPI.

These issues only arrived once dynamic tick was introduced. Prior to dynamic
tick, clock IPIs were only sent from the clock owner (P0) to the other cores
when the clock was not per-processor.

Besides the extra IPIs being sent around and the extra trips to the
hardware, this never surfaced as a bug until the introduction of absolute
timers. As absolute timers get rearmed by the acknowledge interrupt handler,
it is incorrect to rearm them when a software generated clock IPI arrives.
This causes then next due time to be scheduled before the current due time
expires, leaving unexpected gaps in the real hardware-generated clock
interrupts.