Fixed absolute timer interrupt bursts.
When an absolute timer gets rearmed during acknowledge interrupt, it adds
the period to the previous due time. This works great up until the core
gets frozen for a debug break or print - or perhaps a power state in the
future. When this happens, the next calculated due time may be in the past.
Setting such a due time causes the absolute timer's interrupt to
immediately fire and it will keep rearming to a due time in the past until
it has interrupted enough times for the periodic addition to catch up with
the current time. This creats a temporary interrupt storm.

So, this change takes the head of reading the current timer value before
rearming the timer. If the current time is ahead of the due time, the next
interrupt is scheduled a period out in the future.