Added GT support for the Raspberry Pi 2 & 3.
This change adds the necessary changes to support the ARM Generic Timer on
the Raspbeery Pi 2 & 3. This includes core timer clock initialization in the
firmware, a GT virtual offset reset to 0 on each core while in HYP mode, a
GTDT ACPI table passed to the kernel by the firmware, and updates to the
BCM2709 interrupt controller to handle the per-processor interrupts for the
core timer (i.e. the ARM Generic Timer).