Debugger resilience towards crossfired messages.
This change tries to make the debugger a bit more resilient to cases where
the client and the target talk over each other. Specifically, when memory
profiling is enabled, and a break request was sent just as profiling data
was being sent, the debugger took over a minute to resynchronize and
connect.

The main change is to swallow the payload in the client if a valid header
was received. The biggest set of problems came from misinterpreting data
bytes as synchronization bytes.

Additionally, ignore failure in Kd, and only look at the break in request
flag for prints and profiling data.