Fixed usermode debugger single step handling.
Both the single step and range step debugger commands set the debugee into
single step mode. Once that results in a trap, the process should be taken
out of single step mode. This was broken due to the command state being set
to invalid in order to protect against spurious signals. This change adds a
specific previous command field that is checked on each trap.