Add itimer support in the kernel.
Add support for getitimer and setitimer in the kernel, which enables
setting a wall clock timer, a timer based on user cycle execution, and a
timer based on user+kernel cycle execution. Nginx uses this.

I'm holding back on committing the C library portion of this so that the
automated builds can rev once. If I remember correctly build failures
occur if the new libc source depends on a function that is not in the old
libminocaos.