Add basic SIGIO support.
This adds support for F_GETOWN, F_SETOWN, and FIOASYNC (ioctl). It
currently sends SIGIO differently than other OSes. We send SIGIO only
when there's a change in IO state (ie was not ready to read and now is).
Other OSes send SIGIO whenever there might be a change (ie for each new
data that is available to be read). We'll see if this works out or not.