Fixed stddef.h to define NULL on first include.
stddef.h is taken from GCC and only defines ptrdiff_t, size_t, wchar_t, and
NULL if the appropriate __need_* is set or if none of the __need_* macros
are defined. libcbase.h was setting __need_* for all except NULL, causing
the others to be defined, but not NULL.  This was likely due to wint_t's
behavior where it does not get defined unless __need_wint_t is set.

This change limits the use of __need_* to the __need_wint_t case in wchar.h.
Now every other use of stddef.h will go through and define ptrdiff_t,
size_t, wchar_t, and NULL.