Add temporary shims for setjmp and longjmp.
The headers in the automated builds are always one generation behind,
which means they'll be pointing to functions for setjmp and longjmp.
Create temporary shim functions so that these link properly. In the next
build iteration, the macros to _setjmp and _longjmp will kick in, and
these can be removed.