Move KPROCESS.ControllingTerminal to be an IO_HANDLE.
To facilitate an implementation of IoOpenControllingTerminal that returns
an IO handle, the controlling terminal pointer in each process was
converted to an IO handle pointer. It now lives in each process rather
than just the session leader, so that a non-session leader can
lose its controlling terminal, and so that a session leader can gain a
different controlling terminal without the rest of the session having it.

The tricky part is that each process does not have a reference to this IO
handle. It's added when a session leader opens a controlling terminal,
copied on fork, and cleared in everyone when the session leader closes
that terminal. No one should touch the controlling terminal without the
terminal list lock held, because the session leader might close it at any
time.

The one exception is PspCopyProcess, which is a little tricky.
Disassociate will have to iterate over all the processes to remove the
controlling terminal from each, which requires having the process list
lock held. When forking, the controlling terminal and session ID are set
before adding the process to the global list. Then, a double check is done
to see if the controlling terminal was cleared from the parent in the
meantime, and is also cleared from the new child.