Added special device to open current terminal.
Wired up this special device to /dev/tty so that applications can open
their controlling terminal as they expect. One part of this implementation
that I don't love is that special has to store the original I/O handle it
opened and close it when its own handle is closed. Ideally we'd have some
sort of seamless mechanism for magically opening a different
file object. But the trouble is the opens and closes don't match without
some sort of reference counting on the open itself (rather than the
IO_HANDLE).