Fix paging I/O bugs caused by IO_HANDLE.FileObject.
An I/O handle is not always an IO_HANDLE if it's a paging I/O handle.
Watch out for that in the couple of places we were using IoGetPathPoint
before. Ideally we would separate out the paging I/O path into its own
area. The only problem seems to be FAT opens the disk in either paging or
non-paging mode, and needs and IO handle back either way.