Clean up C library terminal functions for /dev/tty.
The terminal ioctl wrappers don't need to call isatty beforehand: those
IOCTLs will fail if the descriptor is not a terminal.

Use tcgetattr as a test for isatty. The old test of using something like
fstat is not good because /dev/tty's fstat will not report itself as a
terminal slave (since it operates on the original file object opened
rather than the one doing the IO).