Added function to convert timespec to timeout in ms.
The OS API functions all take timeouts in milliseconds. A few C library
routines take timeouts in timespec values. The converstion and error cases
are the same across all the routines. This change adds a helper routine to
centralize converting a timespec value into a millisecond timeout,
accounting for timespec values that are beyond the maximum supported timeout
of ~49 days.