Add resource limit scaffolding.
Added the storage and system calls to maintain resource limits in the
kernel for each thread. The only limit that is currently implemented is
the stack size.

This is in preparation for moving memory mapped regions up to the top, so
that brk can be implemented, so that emacs undump works (ish).