Moved PROCESS_ENVIRONMENT block to the user stack.
Before this change we would just map some region of memory and put the
environment block there. This won't work so well for emacs undump, as the
mapped region is usually a very low address, and undump would end up
saving that block. Additionally, applications are probably used to
RLIMIT_STACK capping both the stack and the arguments. A mild side benefit
is the environment doesn't need to round up to a page, as the stack
simply starts right beyond it.

Most of the work in this change came from moving things around to satisfy
the conditions that 1) when the user process environment block is created
the user stack must already be set up, and 2) The stack pointer register
needs to point just beyond the environment block in this case.

Also removed an unused THREAD_FLAG_ADD_REFERENCE.