Improved Chalk string constant and method call lookup.
This change improves the CK_SYMBOL_TABLE structure into a generic
CK_STRING_TABLE structure that can store a set of strings for both fast
indexing and fast lookup by string. It does this by maintaining both an
array and a dict.

Created a new opcode for string constants, and moved string constants to a
module-level string table. Multiple instances of the same string constant
within a module now reuse the same constant index.

Also made the operand to call instructions into an index into this module
level string table. Signatures like __add@1 are now inserted as strings
into this table. Method lookup will need to index into this table to get a
string (fast), then do a dict lookup on the class object (slightly less
fast). This solves the issue of having a giant global array of signature
strings throughout the whole VM, creating too-sparse arrays in each class.
It also alleviates problems with trying to save the compiled bytecode for
a module into a file, as those global method signature numbers were
dependent on the order in which modules were loaded.

I spent some time thinking about trying to separate out the dot (.)
operator from a function call, so that you could do a = myvar.myfunc,
a(1, 2, 3). The problem is that myvar.myfunc would have to evaluate to a
new object that described a function bound to an instance. This is a new
object allocation each time you do myvar.myfunc. Now some sort of internal
dict like __attrs could alleviate that for class instances, but there's
nowhere to cache that information for primitives like integers. The
overhead of allocating a new bound function instance for every operator
seemed like too much of a perf hit, so dot continues to be tied
grammatically to a function call. I'm planning on adding a -> operator
that converts to calls like __getAttr and __setAttr that can be
overridden, so that will hopefully alleviate the functionality tradeoffs
with . made for performance.