Specify min/max address when allocating from Mdl.
Allow memory descriptor list allocation to take a minimum and maximum
viable address to allocate from. This is in preparation for reserving an
area for the user mode stack, and then limiting mmap from allocating
there.

This is needed to in order to move mmap regions to the top of VA space, in
order to cleanly implement brk, so that emacs can use its own heap and
(hopefully) successfully implement their horrible undump hack.