Added __STDC_ISO_10646__ to libcbase.h.
MySQL requires __STDC_ISO_10646__ to be defined. This is set to a yyyymmL
value that corresponds to the year and month of the ISO/IEC 10646 standard
whose required Unicode character set is supported by the C Library. In this
case, support requires that the short identifier of every character in the
Unicode standard set has the same value when stored in a wchar_t.
Essentially, if a wchar_t is 32-bits, then all the characters are supported,
as they only need 21-bits (to date).