Changed mmap to return high addresses.
Limited the region where MmSysMapOrUnmapMemory will return memory. This
saves a region for the stack, with some padding for future expansion. Now
memory mapped regions return high (and descending) addresses. The heap and
all dynamic libraries should now be up near the top of user mode memory.

This is so that when brk is implemented it won't collide with mmap'ed
regions at the bottom.