Fix detached pthread memory leak.
pthread_exit was failing to deallocate PTHREAD.OsData in the case where
the thread was detached. This became apparent with the new mmap changes as
mmap now prefers the highest address (and just starts with the biggest
chunk), rather than the exact best fit. So the pthread_detach perf test
was burning through all its VA and failing.