Taught libminocaos.so to relocate itself.
In order to have all mmap calls (including image loads) default to the top
of memory, libminocaos has to be able to relocate itself. This turned out
to be not that hard, it just involves initializing and wiring together the
right structures.

The one bummer is performance seems to have suffered, in the form of more
page faults. Perf tweaks to come.