Create destination path entry in IoRename for perf.
Do a little more work in IoRename to create the destination path entry
that is known to exist. This saves the underlying file object from being
destroyed, which also saves the painful cost of having to look it up again
in the file system.

Also fixed a minor bug where rename the file to itself was destroying the
path entry, forcing a lookup next time. This was highlighted by the bug in
the perf rename test.