Only write portion of FAT window that is dirty.
A change back in May halved the result of the create perf test. In order
to limit VA fragmentation, the FAT cache window size was increased from 2K
to 128K. Unfortunately this meant that whenever a portion of the FAT
becomes dirty, the minimum size to write out is now 128K. This change
converts the Dirty boolean into a min/max region, so that only the portion
of each window that is dirty needs to be written out.