Track the boundaries of an image section mapped region.
The user mode stack defaults to 8MB, but usually nothing close to that
amount is ever touched and mapped in. A lot of time is wasted getting PTEs
for regions that are never touched during unmap.

This change tracks the min and max addresses that have ever been mapped
for an image section. It can use these boundaries to save on loop
iterations during unmap and copy.