Truncate should make it to the file system.
Awhile back we changed writes to hit the file system to avoid
overcommitting storage space. Truncate should get the same treatement.
This required teaching FAT how to grow a file, which it sort of already
had been doing in the old version of write (where it might get surprised
by a new large file size).

This was found by an assert in the shared memory object code fired during
the mmaptest.