Fix bug causing page tables not to get created.
The stack image section spans two page directories, but usually only
touches the last page or so. On fork, page tables are created but not set
to present in the child process. With the new Min/MaxTouched variables in
image sections, MmpCopyAndChangeSectionMappings is called for only that
small sliver at the end, so the other page tables in the stack section
never get set to present. Then, if that new child forks,
MmpPreallocatePageTables will skip over those page tables, since it looks
only at the present bit of the PDE, not the Entry.

Make preallocate observe the entry rather than the present bit, so
non-present but still necessary page tables are also copied. The existing
code already watches out for needing to zero a page table that exists but
is uninitialized.