Changed int32_t and uint32_t to use int.
int32_t and uint32_t are supposed to be exactly 32-bits according to the
POSIX specification. They were previously defined to long and unsigned long,
making them 64-bit values on 64-bit architectures. This change converts them
to int and unsigned int, which are consitently 32-bit values on both 32-bit
and 64-bit architectures.