select should return error mask in read/write masks.
If the error mask is not supplied to select, but it happens, then the read
and/or write masks should be set to reflect it. Cmake was selecting a
pipe, which the kernel would return 1 descriptor ready from, but only
passed in the read bitmask. Since the bit wasn't set on the return from
select, it would just loop back and call select again, never making
progress.

Also lined up the pollfd and POLL_DESCRIPTOR structures so that no copies
have to occur. There is going to be an issue in 64-bit when pollfd
continues to have an int and the kernel's HANDLE will be 64 bits wide. Not
sure exactly what to do there.