Added RTLD_DEFAULT and RTLD_NEXT for dlsym.
When RTLD_DEFAULT is supplied to dlsym as the library handle, a global
symbol search will be performed starting with the executable. When
RTLD_NEXT is supplied to dlsym as the library handle, the symbol search
will be performed starting with the first library loaded after the primary
executable.

This change also adds support for dlsym resolving TLS symbols and sorted
out some bugs regarding the symbol search order.