Fix getaddrinfo assert for own hostname.
Emacs was using gethostname and then calling getaddrinfo with that
hostname hoping for the fully qualified domain name. This was asserting
because the local address information didn't have a name filled out (it
assumed you'd never go from a name to a name).

gethostname was also incorrectly returning the FQDN. Add getdomainname to
round out the functions, and an internal function to get the FQDN.