Adjust overhang behavior in videocon.
Turns out the original overhang behavior in the VT100 is slightly
different than what we were doing. If you put a character in column 80,
they would fill it in but leave the cursor on that same position with a
flag set. A printable character would end up in column 1 of the next line.
A crlf however would also only advance to the next line, rather than going
down two lines.

We were previously honoring overhang but advancing the cursor to the next
line. This caused problems when tmux would go to the last line and write
exactly Width characters, because it caused a screen scroll.

Bash actually adjusts for the weird overhang by printing an extra space if
you fill in the last column in the line.