Rework controlling terminal pointer.
The terminal I/O checks for "am I the foreground process group" should
only be done if the terminal is your controlling terminal. Unfortunately
determining whether a given terminal is your controlling terminal is
tricky when the KPROCESS.ControllingTerminal pointer is an IO_HANDLE
(since opening /dev/tty may give you a different IO_HANDLE to the same
terminal).

So I reworked KPROCESS.ControllingTerminal to store a file object. In
order to not have to expose the file object definition out to drivers, I had
the Open IRP pass down the IO_HANDLE it was creating, and then
IoOpenControllingTerminal overwrites the file object in that handle by
taking in an existing IO_HANDLE.

This also modified the behavior of the session leader. The old code sent
the hangup signal when the session leader closed the IO handle that caused
the controlling terminal assignment. Now the session leader only fully
disassociates when its process dies. I think this is closer to correct.

Now (in a future change) I can determine whether a given terminal is my
controlling terminal by comparing pointers: no lock needed. This should
make tmux happier.