Bug fixes in MmpCopyAndChangeSectionMappings.
The continue early in the outer loop was causing CurrentVirtual to get out
of sync with the first level index. This caused the table index start to
get way out of line, which caused zeroing way beyond the end of the page
table page. This fix keeps CurrentVirtual synced up with the loop.

The other more minor fix is to adjust the loop boundaries. It used to go
to <= TableIndexEnd. This is bad because if the caller had a VirtualEnd of
KERNEL_VA_START, the routine would go start changing kernel mappings. This
never happened because the shared user data page always acted as a buffer,
but relying on that is dangerous. The fixed version calculates the
proper ending directory index, and goes strictly less then that.