Minor bug fixes to videocon.
Tmux exposed a couple bugs. One is that if there is a bottom margin (ie
scrolling region is set above the bottom of the screen) the last line of
the screen was not properly getting preserved.

The other bug was that if the cursor is set outside the scroll region,
newline should still move the cursor down until it hits the bottom of the
screen. Our old behavior just kept the cursor at its current line if it
was outside the scroll region.