Fixed image section isolation.
Image section isolation causes an image section to map a copy of its page
into each of its children. The contents of that page may be identical to
the original source (backing image / page cache) or the section may have
dirtied the page before creating the child via fork.

The latter case was causing problems when paging out from the child. The
page would not be seen as dirty (the PTE bit is not set) and would not be
written to the page file. On subsequent page-ins, the page would be drawn
from the original backing image / page cache, mapping inaccurate data.

There was also an issue with the former case where the page was not
different from the original source. Page out would toss the page and page
in would map the correct page back into place. Great. Except if the
section were page cache backed, the page cache would not know that the
section (the child) is inheriting from the page cache. As the parent
isolated its children, the section would not have the bit set in the
inherit bitmap. But since the inherit bitmap was overloaded with two
meanigns (inherit from parent and inherit from page cache), the page cache
would miss unmapping this child section and free the page while the child
still had it mapped. Complicated for sure, but tests hit this case.

The fix is to no longer overload the inherit bitmap. That now always
indicates whether or not a section inherits from its parent. The dirty
page bitmap is now used to indicate whether or not a page differs from
the original source. The dirty state is passed down to a child on fork.
This fixes the first case outlined above as page out would see that the
isolated child is actually dirty (from the parents pre-fork write) and
write the page out to the page file. It also fixes the second case
outlined above as the page cache would see that the child section is not
dirty (i.e. it maps the original source) and would proceed to unmap the
section before freeing the page cache entry.

There is an included optimization for page cache backed sections. When a
section is not dirty and goes to isolate its children, the children can
continue to map the page cache pages as their dirty page bitmaps now hold
the answer to whether or not the section maps the page cache page at that
offset.