Optimized dying image section isolation.
There is no reason to isolate a dying image section from its parent by
allocating a new page, copying the parent's content into said page and
mapping it in the section. The section is just going to turn around to
unmap and free the page.

This change makes dying image sections only isolate their children and not
themselves.