Fix a couple pthreads bugs.
One bug was causing an infinite AtomicCompareExchange loop in the case
that the first CompareExchange failed because the state never got updated
to what it actually is.

Another bug was causing pthread_once to go down for a wait indefinitely
if the completion went through at just the right time.