Set dirty bitmap in children on section isolation.
Rather than relying on a well-hidden x86 code path that marks all pages
mapped in another process dirty, this change puts the dirty propogation in
the common code path. When isolating a section, if the section's page is
determined to be dirty all of its children's copies of that page will be
noted as dirty in the children.

This not only depends on the sections dirty bitmap, but the actually page
table entry bits; the section may have been mapped writable before the
fork that created the child. The dirty bitmap is never set in that case.