Free up eax in sysenter mechanism.
In preparation for trying to save very little during system calls, I moved
the stack used by sysenter (aka the double fault stack) to also coincide
with the main TSS. This way, sysenter can get to its stack with a single
SS: dereference (rather than needing a DS: dereference, which requires a
register to load DS). This freed up eax, which can now be used as the
system call number.

This adds a small amount of complexity as the main TSS must be on the
double fault stack. For processor 0, this means reinitializing the main
TSS. For the others, it's just kind of a funny looking swaparoo.

This is ultimately towards delivering signals on the user mode stack, to
support nested signals.