Added si_fd to siginfo_t.
The bulk of the work was already done. This just pipes the asynchronous I/O
receiver's file descriptor back to user mode. The descriptor is saved when
an I/O handle signs up for asynchronous I/O. There can really be only one
descriptor per process for an I/O state, which is associated with a file
object. The I/O signal sending code will just put the first descriptor found
for the signaling process in the parameters.