Fixed issues in select/pselect.
Cmake was still hanging because it was selecting a single pipe for read
(only). The bit wasn't getting set because the only event coming back was
POLL_EVENT_DISCONNECT, but the error part that was supposed to force
IN|OUT on was only checking for POLL_EVENT_ERROR, rather than
POLL_ERROR_EVENTS.

Also noticed that POSIX mentions no bitmasks should be changed on error.
So do a quick scan for an invalid descriptor before modifying the
bitmasks. This one hadn't caused a problem yet, but might if an app was
lazy about refreshing their bitmasks.