Use a different hint for incomplete trap frame on ARM.
Qemu was allowing the reserved bit 20 of the CPSR to actually be set,
causing failures on things like fork since the trap frame looked
incomplete but wasn't. Restoring the bit was starting to look like more
work than just setting another value. So I moved the hint to be
TRAP_FRAME.ExceptionCpsr == -1. This should never happen accidentally
because bits 20-23 at least are reserved in the CPSR and should be zero.