Move struct tm.tm_zone to be a char *.
Things like Ruby and Guile were starting to choke on the fact that tm_zone
was an array rather than a pointer. This changes moves
CALENDAR_TIME.TimeZone to be a pointer as well. Rtl caches and reuses
strings it hands out in these structures. The C library uses these
pointers directly.