Fixed mmap suggested address behavior.
If a non-zero address is supplied to mmap, the system should make a best
effort to use that address if it is free. The file section mapping routine
was clobbering the suggested address.

Also, when an address is suggested, there needs to be a fallback strategy.
Rather than storing two strategies in the VA request struction, the system
now assumes that a non-zero address is a preferred or fixed address request
and treates the strategy as the fallback strategy in the case where the
suggested address is not free.