Added posix_memalign to the C Library.
Because memory allocated via posix_memalign can be released using free(),
the required functionality must be implemented in the lower level RTL heap
library. The algorithm is to simply allocate a large enough chunk to
support alignment, find an aligned range in that chunk and free the head
tail chunks back to the heap.

This change includes an update to the RTL heap tests to mix in some calls
to RtlHeapAlignedAllocate amongst regular allocate, reallocate, and free
calls.