Add dl_iterate_phdr, link_map, and r_debug stuctures.
Boehm GC needs to keep an eye on the dynamic libraries loaded in order to
accurately determine GC roots. This change modifies the LOADED_IMAGE
structure so that link_map and r_debug can overlay on it.

Also reworked what image information is kept in the debugger. All that's
really needed is the lowest loaded address, as the rest can be computed
using the image itself.