Fixed bug that stranded new threads without a signal.
If a signal (like kill) was sent to a thread that had not yet run for the
first time, it would get a chance to run normally before the signal was
delivered. The Boehm GC threadkey_test was in the middle of creating a
thread when it exited. The thread's start mutex was never unlocked, and it
blocked immediately (with signal pending state set to unknown), so it
ended up being stuck until another signal (like SIGINT) was dispatched.

The solution is to check for signals before starting a user mode thread
for the first time.