Block all signals while creating a thread.
It occurred to me that a signal coming in too early in a newly created
thread could do some bad things, like fail to find itself in the thread
list or reach through an uninitialized ClCurrentThread.

Block all signals until the thread is ready to accept them. Stop, kill,
and continue cannot be blocked, but they also don't require any action on
the part of the thread.