Force important signals sent by the kernel.
Usermode should not be allowed to block signals like SIGSEGV, SIGBUS,
SIGILL, and SIGFPE when sent by the kernel. Force those signals to cut
through the blocked and ignored masks.

This also means that the ignore mask should only be observed when a signal
is queued. By the time a signal is dequeued, it's assumed that signal's
not ignored (or wasn't when it was queued). This is because there's
currently no way of knowing at dequeue time whether a signal was forced or
not. It was just double work for no reason.