Track image real file paths to dedupe library opens.
Guile (via libltdl) was getting confused because it would try to dlopen a
symlink to another library. Our loader would load that as a new library
rather than detecting a previously opened library.

This change rearranges the open file path of the image loader. This will
make the DT_RPATH and DT_RUNPATH paths more functionally correct, and
allows the final path of the opened file to come out at the end.

Unfortunately OsGetRealPath has to make a system call for each path
component. This is awfully heavy. I think I can do this by just comparing
device and file numbers instead (though that won't work for the images
loaded by the kernel). Also OsGetRealPath seems a bit too C library-ish
for libminocaos anyway.

New commit undoing some of this stuff coming next, but I wanted to save it
in case we need to revisit this method.