Fix terminal deadlock.
The locking order between the terminal input and output locks was meant to
be output followed by input. IopTerminalMasterWrite was violating the order.
In various locations (subroutines) it was dropping and reacquiring the
output lock while maintaining hold of the input lock.

This change switches the lock ordering to be input followed by output. It
also required some cleanup in IopTerminalMasterWrite to not release the
input lock in certain locations. Removing those releases simplify the lock
ordering and actually fix a theoretical race condition between two master
writers.